requirements.txt
grpcio-tools ~= 1.30
flask-sqlalchemy
PyMySQL
grpc-interceptor ~= 0.12.0
grpcio-tools ~= 1.30
pytest ~= 5.4